move CommandChannel to util mod
